refactor(uuid): make v7 APIs explicitly unstable
