+++ uuid/deno.json
@@ -9,6 +9,6 @@
     "./v3": "./v3.ts",
     "./v4": "./v4.ts",
     "./v5": "./v5.ts",
-    "./v7": "./v7.ts"
+    "./unstable-v7": "./unstable_v7.ts"
   }
 }
